Move scored_graph code into it's own module. Fix some bugs with the recipe db.
